feat: 파일에서 redis에 저장으로 변경 fix: 기존의uint32_t의 자료형인 timestamp를 uint64로 변경 밀리세컨드 단위의 저장을 위함

+++ testChatServer/ChatMessageWrapper.cpp
@@ -8,7 +8,7 @@ ChatMessageWrapper::ChatMessageWrapper()
 {
 }
 
-ChatMessageWrapper::ChatMessageWrapper(const std::string &userName, uint32_t timestamp, chat::DataType dataType)
+ChatMessageWrapper::ChatMessageWrapper(const std::string &userName, uint64_t timestamp, chat::DataType dataType)
 {
     message_.set_user_name(userName);
     message_.set_timestamp(timestamp);
@@ -53,7 +53,7 @@ std::string ChatMessageWrapper::getUserName() const
     return message_.user_name();
 }
 
-uint32_t ChatMessageWrapper::getTimestamp() const
+uint64_t ChatMessageWrapper::getTimestamp() const
 {
     return message_.timestamp();
 }
